move stats components to library package
